requirements.txt
python-telegram-bot[job-queue]~=21.0
openai~=1.60
python-dotenv~=1.0
aiosqlite~=0.20
